// valorant-api/models/PremierConferences.ts
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export enum PremierConferences {
    EuCentralEast = 'EU_CENTRAL_EAST',
    EuWest = 'EU_WEST',
    EuMiddleEast = 'EU_MIDDLE_EAST',
    EuTurkey = 'EU_TURKEY',
    NaUsEast = 'NA_US_EAST',
    NaUsWest = 'NA_US_WEST',
    LatamNorth = 'LATAM_NORTH',
    LatamSouth = 'LATAM_SOUTH',
    BrBrazil = 'BR_BRAZIL',
    KrKorea = 'KR_KOREA',
    ApAsia = 'AP_ASIA',
    ApJapan = 'AP_JAPAN',
    ApOceania = 'AP_OCEANIA',
    ApSouthAsia = 'AP_SOUTH_ASIA'
}